requirements.txt
boto3
langchain-aws
langchain-core
python-dotenv
langchain-community
faiss-cpu
langchain
google-generativeai
langchain-google-genai
anthropic
pinecone